refactor(tangle-dapp): move ensureError to tangle-shared-ui
