+++ apps/tangle-dapp/app/bridge/hooks/useBalance.ts
@@ -3,6 +3,7 @@
 import { isAddress } from '@polkadot/util-crypto';
 import { ChainConfig } from '@webb-tools/dapp-config/chains/chain-config.interface';
 import { AddressType } from '@webb-tools/dapp-config/types';
+import ensureError from '@webb-tools/tangle-shared-ui/utils/ensureError';
 import { useWebbUI } from '@webb-tools/webb-ui-components/hooks/useWebbUI';
 import Decimal from 'decimal.js';
 import { useCallback, useEffect, useMemo } from 'react';
@@ -10,7 +11,6 @@ import useSWR from 'swr';
 
 import { useBridge } from '../../../context/BridgeContext';
 import useActiveAccountAddress from '../../../hooks/useActiveAccountAddress';
-import ensureError from '../../../utils/ensureError';
 import { isEvmAddress } from '../../../utils/isEvmAddress';
 import {
   getEvmContractBalance,
