+++ apps/tangle-dapp/components/PercentageInput.tsx
@@ -1,9 +1,9 @@
 import BackspaceDeleteFillIcon from '@webb-tools/icons/BackspaceDeleteFillIcon';
+import ensureError from '@webb-tools/tangle-shared-ui/utils/ensureError';
 import { Button, Input } from '@webb-tools/webb-ui-components';
 import { FC, useCallback } from 'react';
 
 import useCustomInputValue from '../hooks/useCustomInputValue';
-import ensureError from '../utils/ensureError';
 import InputWrapper, { InputWrapperProps } from './InputWrapper';
 
 export type PercentageInputProps = {
