+++ apps/tangle-dapp/hooks/useApiRx.ts
@@ -2,10 +2,10 @@
 
 import { ApiRx } from '@polkadot/api';
 import useNetworkStore from '@webb-tools/tangle-shared-ui/context/useNetworkStore';
+import ensureError from '@webb-tools/tangle-shared-ui/utils/ensureError';
 import { useCallback, useEffect, useState } from 'react';
 import { catchError, Observable } from 'rxjs';
 
-import ensureError from '../utils/ensureError';
 import { getApiRx } from '../utils/polkadot';
 import usePromise from './usePromise';
